Refactor other cluster functions
