+++ src/model/services/mapi/capiv1alpha3/getClusterInfraRef.ts
@@ -1,50 +1,31 @@
+import { IOAuth2Provider } from 'lib/OAuth2/OAuth2';
 import { IHttpClient } from 'model/clients/HttpClient';
-import { LoggedInUserTypes } from 'stores/main/types';
 
 import * as capzv1alpha3 from '../capzv1alpha3';
+import { getResource } from '../generic/getResource';
 import { ICluster } from './';
 
 export function getClusterInfraRef(
   client: IHttpClient,
-  user: ILoggedInUser,
-  cluster: ICluster
+  auth: IOAuth2Provider,
+  url: string
 ) {
-  return async () => {
-    const { infrastructureRef } = cluster.spec;
-    if (!infrastructureRef) {
-      return Promise.reject(
-        new Error('There is no infrastructure reference defined.')
-      );
-    }
-
-    switch (infrastructureRef.kind) {
-      case capzv1alpha3.AzureCluster:
-        return capzv1alpha3.getAzureClusterByName(
-          client,
-          user,
-          cluster.metadata.namespace!,
-          infrastructureRef.name
-        )();
-
-      default:
-        return Promise.reject(new Error('Unsupported provider.'));
-    }
-  };
+  return getResource<capzv1alpha3.IAzureCluster>(client, auth, url);
 }
 
 export function getClusterInfraRefKey(
-  user: ILoggedInUser | null,
+  client: IHttpClient,
+  auth: IOAuth2Provider,
   cluster: ICluster
-): string | null {
-  if (!user || user.type !== LoggedInUserTypes.MAPI) return null;
-
+) {
   const { infrastructureRef } = cluster.spec;
   if (!infrastructureRef) return null;
 
   switch (infrastructureRef.kind) {
     case capzv1alpha3.AzureCluster:
-      return capzv1alpha3.getAzureClusterByNameKey(
-        user,
+      return capzv1alpha3.getAzureClusterKey(
+        client,
+        auth,
         cluster.metadata.namespace!,
         infrastructureRef.name
       );
