+++ src/model/services/mapi/capzv1alpha3/index.ts
@@ -1,2 +1,2 @@
 export * from './types';
-export * from './getAzureClusterByName';
+export * from './getAzureCluster';
